refactor(ui): move the UserButton contract into user-button.types
The data contract, mode flags, and menu item shapes are what the view and
a controller agree on, so they sit in a module neither one owns. Annotates
the exported parts' return types: without one TS infers the global
JSX.Element, which Emotion augments, so the signature published to apps
read EmotionJSX.Element.
