+++ packages/ui/src/mosaic/user-button/user-button.view.tsx
@@ -2,7 +2,7 @@
 
 import type { PopoverProps } from '@clerk/headless/popover';
 import * as stylex from '@stylexjs/stylex';
-import type { ReactNode } from 'react';
+import type { ReactElement, ReactNode } from 'react';
 import React from 'react';
 
 import type { AvatarProps } from '../components/avatar';
@@ -20,123 +20,24 @@ import { Spinner } from '../components/spinner';
 import { truncationStyles } from '../components/typography.styles';
 import type { IconName } from '../icons/registry';
 import { fontWeightVars } from '../tokens.stylex';
-import type { UserButtonMenuItem, UserButtonMenuItemId } from './user-button.menu';
 import { arrangeMenuRows } from './user-button.menu';
 import { styles, triggerShapes } from './user-button.styles';
+import type {
+  UserButtonBusyState,
+  UserButtonCallbacks,
+  UserButtonData,
+  UserButtonMembership,
+  UserButtonMenuItemId,
+  UserButtonMenuProps,
+  UserButtonMode,
+  UserButtonModePriority,
+  UserButtonModeProps,
+  UserButtonSession,
+} from './user-button.types';
 
-// ─── Data contract ──────────────────────────────────────────────────────────
-// Session-backed, discriminated resource rows. Intended to be 1:1 with a future
-// `useUserButtonController()` output so the controller is a drop-in follow-up.
-
-export interface UserButtonSession {
-  sessionId: string;
-  name: string;
-  /** Whatever the account is addressed by: username, email, phone, or wallet. */
-  identifier: string;
-  imageUrl?: string;
-}
-
-export interface UserButtonMembership {
-  kind: 'membership';
-  organizationId: string;
-  name: string;
-  imageUrl?: string;
-  membersCount?: number;
-  planLabel?: string;
-}
-
-export interface UserButtonSuggestion {
-  kind: 'suggestion';
-  id: string;
-  organizationId: string;
-  name: string;
-  imageUrl?: string;
-  /** `accepted` is awaiting approval, so it lists but cannot be joined again. */
-  status: 'pending' | 'accepted';
-}
-
-export interface UserButtonInvitation {
-  kind: 'invitation';
-  id: string;
-  organizationId: string;
-  organizationName: string;
-  imageUrl?: string;
-  /** `accepted` is already a workspace, so it lists as one rather than offering to be accepted. */
-  status: 'pending' | 'accepted';
-}
-
-/** Feeds one more page into a list as its foot scrolls into view. */
-export interface UserButtonPaging {
-  ref: (element: HTMLElement | null) => void;
-  hasMore: boolean;
-}
-
-export interface UserButtonData {
-  activeSession: UserButtonSession;
-  /**
-   * The active organization, described whole rather than found in `memberships`, so the surface
-   * names it while the list it belongs to is still loading. `null` => the personal workspace.
-   */
-  activeOrganization: UserButtonMembership | null;
-  /**
-   * Explicit; do not derive from `memberships.length`. Answered before the lists are fetched, so
-   * the surface knows whether to carry a workspace section at all without waiting on them.
-   */
-  hasOrganizations: boolean;
-  /**
-   * The account has no workspace of its own to return to, so the organizations are all there is.
-   * Withholds the personal row rather than standing it down: this is not a switch that is
-   * momentarily unavailable, it is a workspace that does not exist here.
-   */
-  hidePersonal?: boolean;
-  /**
-   * A first page is still in flight, so the workspace rows stand in as one placeholder rather than
-   * appearing a list at a time.
-   */
-  organizationsLoading?: boolean;
-  memberships: UserButtonMembership[];
-  suggestions: UserButtonSuggestion[];
-  invitations: UserButtonInvitation[];
-  paging?: UserButtonPaging;
-  /**
-   * The other signed-in accounts. Only sessions: an account's organizations are scoped to the
-   * session that fetches them, so they are unknowable until it is the active one.
-   */
-  additionalSessions: UserButtonSession[];
-}
-
-/** All optional. An unhandled action hides (or de-activates) the affordance it drives. */
-export interface UserButtonCallbacks {
-  /**
-   * Acts on the active account; another account's organizations are unreachable until you switch.
-   * `null` selects the personal workspace, which is how an account leaves an organization.
-   */
-  onSelectOrganization?: (organizationId: string | null) => void;
-  onAcceptSuggestion?: (suggestionId: string) => void;
-  onAcceptInvitation?: (invitationId: string) => void;
-  onSwitchSession?: (sessionId: string) => void;
-  onSignOutSession?: (sessionId: string) => void;
-  onSignOutAll?: () => void;
-  onManageOrganization?: () => void;
-  onInviteMembers?: () => void;
-  onManageAccount?: () => void;
-  onCreateOrganization?: () => void;
-  onAddAccount?: () => void;
-}
-
-/**
- * Which switchers the surface carries. `combined` is both; `orgs` is an organization switcher with
- * no account rows; `user` is an account switcher that never shows an organization, even when one
- * is active.
- */
-export type UserButtonMode = 'combined' | 'orgs' | 'user';
-
-/**
- * Which of the two switchers a `combined` surface leads with: the one named in the trigger and
- * headed in the popup. Both are still listed either way. The single-purpose modes have only one
- * thing to lead with, so they ignore it.
- */
-export type UserButtonModePriority = 'organizations' | 'user';
+// The data contract, the mode flags, and the menu item shapes live in `user-button.types`; they are
+// what the controller and the view agree on, so neither file owns them.
+export type * from './user-button.types';
 
 /**
  * Stable keys naming which affordance owns the single in-flight action. Shared by the connected
@@ -150,26 +51,6 @@ export const userButtonBusyKeys = {
   acceptSuggestion: (suggestionId: string) => `accept-suggestion:${suggestionId}`,
   acceptInvitation: (invitationId: string) => `accept-invitation:${invitationId}`,
 } as const;
-
-/** The app's own actions at the foot of the popup, and the order the foot's rows run in. */
-export interface UserButtonMenuProps {
-  /** Actions and links of your own, added to the foot of the popup ahead of Clerk's own rows. */
-  customMenuItems?: UserButtonMenuItem[];
-  /**
-   * The order the foot's rows run in, by id: a built-in row's id, or a custom item's `id`. Anything
-   * left out follows the rows named here. An id the surface does not carry as a row is ignored,
-   * since which rows the foot has depends on its mode.
-   */
-  menuItemOrder?: (UserButtonMenuItemId | (string & {}))[];
-}
-
-export interface UserButtonBusyState {
-  /**
-   * Key of the single in-flight action (see `userButtonBusyKeys`), or `null`/absent when idle. The
-   * affordance that owns it spins; every other one is disabled so a second action cannot start.
-   */
-  pendingKey?: string | null;
-}
 
 type UserButtonContextValue = UserButtonData &
   UserButtonCallbacks &
@@ -1034,17 +915,8 @@ function Footer() {
 // ─── Public parts ───────────────────────────────────────────────────────────
 
 export interface UserButtonRootProps
-  extends UserButtonData, UserButtonCallbacks, UserButtonBusyState, UserButtonMenuProps {
+  extends UserButtonData, UserButtonCallbacks, UserButtonBusyState, UserButtonMenuProps, UserButtonModeProps {
   children: ReactNode;
-  /** @default 'combined' */
-  mode?: UserButtonMode;
-  /**
-   * Which switcher a `combined` surface leads with in the trigger and the popup's header. The
-   * other one is still listed. Ignored by the single-purpose modes.
-   *
-   * @default 'organizations'
-   */
-  modePriority?: UserButtonModePriority;
   open?: boolean;
   defaultOpen?: boolean;
   onOpenChange?: (open: boolean) => void;
@@ -1057,7 +929,7 @@ export interface UserButtonRootProps
  * the headless `Popover.Root` — it does not keep a second controllable-state copy. Leaves consume
  * the data through context.
  */
-export function UserButtonRoot(props: UserButtonRootProps) {
+export function UserButtonRoot(props: UserButtonRootProps): ReactElement {
   const {
     children,
     mode = 'combined',
@@ -1100,7 +972,10 @@ export interface UserButtonTriggerProps {
 }
 
 /** The trigger: the active workspace's avatar, and what it is called. */
-export function UserButtonTrigger({ renderTriggerLabel = true, renderPlanBadge = true }: UserButtonTriggerProps = {}) {
+export function UserButtonTrigger({
+  renderTriggerLabel = true,
+  renderPlanBadge = true,
+}: UserButtonTriggerProps = {}): ReactElement {
   const data = useUserButtonContext();
   const { name, imageUrl, shape, organization } = leadWorkspace(data);
   const planLabel = renderPlanBadge ? organization?.planLabel : undefined;
@@ -1134,7 +1009,7 @@ export function UserButtonTrigger({ renderTriggerLabel = true, renderPlanBadge =
 export function UserButtonTriggerSkeleton({
   mode = 'combined',
   modePriority = 'organizations',
-}: Pick<UserButtonRootProps, 'mode' | 'modePriority'> = {}) {
+}: UserButtonModeProps = {}): ReactElement {
   const shape = leadsWithOrganization(mode, modePriority) ? 'square' : 'circle';
 
   return (
@@ -1148,7 +1023,7 @@ export function UserButtonTriggerSkeleton({
 }
 
 /** The popover surface: header, workspace list, additional accounts, and footer. */
-export function UserButtonPopup() {
+export function UserButtonPopup(): ReactElement {
   return (
     <Popover.Popup aria-label='Account'>
       {/* The card lays its children out with a row gap; the rows read as one continuous list. */}
@@ -1168,7 +1043,7 @@ export type UserButtonProps = Omit<UserButtonRootProps, 'children'> & UserButton
  * Presentational all-in-one: renders the trigger + popup from a single prop-driven call. The
  * connected, Clerk-backed `UserButton` lives in `user-button.tsx` and wraps this view.
  */
-export function UserButtonView({ renderTriggerLabel, renderPlanBadge, ...root }: UserButtonProps) {
+export function UserButtonView({ renderTriggerLabel, renderPlanBadge, ...root }: UserButtonProps): ReactElement {
   return (
     <UserButtonRoot {...root}>
       <UserButtonTrigger
